Seperating sidebar logic into a component
