+++ client/src/App.js
@@ -13,7 +13,7 @@ import setAuthToken from './utils/setAuthToken';
 
 
 import './assets/images/gt_favicon.png';
-//import './assets/fonts/';
+// import './assets/fonts/';
 import './assets/css/bootstrap.css';
 import './assets/css/font-awesome.css';
 import './assets/css/bootstrap-theme.css';
